+++ src/main/java/com/sbproject/piclocket/repository/PhotoRepository.java
@@ -1,0 +1,9 @@
+package com.sbproject.piclocket.repository;
+
+import com.sbproject.piclocket.model.Photo;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface PhotoRepository extends JpaRepository<Photo, UUID> {
+}